Added authentication {SPRING_SECURITY}

+++ src/main/java/com/springboot/todolist/Service/impl/TaskServiceimpl.java
@@ -2,8 +2,12 @@ package com.springboot.todolist.Service.impl;
 
 import com.springboot.todolist.Dto.TaskDto;
 import com.springboot.todolist.Entity.Task;
+import com.springboot.todolist.Entity.Users;
 import com.springboot.todolist.Reposiytory.TaskRepository;
+import com.springboot.todolist.Reposiytory.UserRepository;
 import com.springboot.todolist.Service.TaskService;
+import lombok.Getter;
+import lombok.Setter;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -13,36 +17,47 @@ import java.util.stream.Collectors;
 public class TaskServiceimpl implements TaskService {
 
     private final TaskRepository taskRepository;
+    private final UserRepository userRepository;
 
-    public TaskServiceimpl(TaskRepository taskRepository) {
+    public TaskServiceimpl(TaskRepository taskRepository, UserRepository userRepository) {
         this.taskRepository = taskRepository;
+        this.userRepository = userRepository;
     }
 
     @Override
-    public TaskDto createTodo(TaskDto taskDto) {
+    public TaskDto createTodo(TaskDto taskDto,Long userId) {
         Task todo = mapToEntity(taskDto);
+        Users user = userRepository.findById(userId)
+                .orElseThrow(() -> new RuntimeException("User not found with id: " + userId));
+        todo.setUser(user);
         Task savedTodo = taskRepository.save(todo);
         return mapToDto(savedTodo);
     }
 
     @Override
-    public List<TaskDto> getAllTask() {
-        List<Task> tasks = taskRepository.findAll();
+    public List<TaskDto> getAllTask(Long userId) {
+        List<Task> tasks = taskRepository.findByUserId(userId);
         return tasks.stream().map(this::mapToDto).collect(Collectors.toList());
     }
 
     @Override
-    public TaskDto getTaskById(Long id) {
+    public TaskDto getTaskById(Long id,Long userId) {
         Task todo = taskRepository.findById(id)
                 .orElseThrow(() -> new RuntimeException("task not found with id: " + id));
+        if (!todo.getUser().getId().equals(userId)) {
+            throw new RuntimeException("Unauthorized! You cannot delete this task.");
+        }
         return mapToDto(todo);
     }
 
     @Override
-    public TaskDto updateTask(Long id, TaskDto taskDto) {
+    public TaskDto updateTask(Long id, TaskDto taskDto,Long userId) {
         Task task = taskRepository.findById(id)
                 .orElseThrow(() ->new RuntimeException("task not found with id " + id ));
         task.setTitle(taskDto.getTitle());
+        if (!task.getUser().getId().equals(userId)) {
+            throw new RuntimeException("Unauthorized! You cannot update this task.");
+        }
         task.setDescription(taskDto.getDescription());
         task.setCompleted(taskDto.isCompleted());
         Task updatedTask = taskRepository.save(task);
@@ -51,9 +66,12 @@ public class TaskServiceimpl implements TaskService {
     }
 
     @Override
-    public void deleteTask(Long id) {
+    public void deleteTask(Long id,Long userId) {
         Task todo = taskRepository.findById(id)
                 .orElseThrow(() -> new RuntimeException("Todo task not found with id: " + id));
+        if (!todo.getUser().getId().equals(userId)) {
+            throw new RuntimeException("Unauthorized! You cannot delete this task.");
+        }
         taskRepository.delete(todo);
 
     }
